Move standard clusters in their own files
