+++ examples/wifi-echo/server/esp32/main/include/LEDWidget.h
@@ -27,7 +27,7 @@
 #include "freertos/FreeRTOS.h"
 #include "freertos/task.h"
 #include "freertos/timers.h"
-#include <lib/datamodel/StandardClusters.h>
+#include <lib/datamodel/ClusterOnOff.h>
 
 class LEDWidget : public chip::DataModel::ClusterOnOff
 {
