+++ src/lib/datamodel/DataModel.am
@@ -32,7 +32,7 @@ CHIP_BUILD_DATAMODEL_HEADER_FILES                              = \
     @top_builddir@/src/lib/datamodel/Cluster.h           \
     @top_builddir@/src/lib/datamodel/ClusterServer.h         \
     @top_builddir@/src/lib/datamodel/EndPoint.h              \
-    @top_builddir@/src/lib/datamodel/StandardAttributes.h    \
-    @top_builddir@/src/lib/datamodel/StandardClusters.h      \
     @top_builddir@/src/lib/datamodel/Value.h                \
+    @top_builddir@/src/lib/datamodel/ClusterBasic.h    \
+    @top_builddir@/src/lib/datamodel/ClusterOnOff.h      \
     $(NULL)
